+++ packages/core/src/pages/account/orders/index.tsx
@@ -23,8 +23,10 @@ import { getMyAccountRedirect } from 'src/utils/myAccountRedirect'
 import { groupOrderStatusByLabel } from 'src/utils/userOrderStatus'
 
 import { MyAccountListOrders } from 'src/components/account/orders/MyAccountListOrders'
-import { extractStatusFromError } from 'src/utils/utilities'
+import { getIsRepresentative } from 'src/sdk/account/getIsRepresentative'
+import storeConfig from '../../../../discovery.config'
 import { validateUser } from 'src/sdk/account/validateUser'
+import { extractStatusFromError } from 'src/utils/utilities'
 
 /* A list of components that can be used in the CMS. */
 const COMPONENTS: Record<string, ComponentType<any>> = {
@@ -53,6 +55,7 @@ export default function ListOrdersPage({
   total,
   perPage,
   filters,
+  isRepresentative,
 }: ListOrdersPageProps) {
   return (
     <RenderSections
@@ -61,7 +64,10 @@ export default function ListOrdersPage({
     >
       <NextSeo noindex nofollow />
 
-      <MyAccountLayout accountName={accountName}>
+      <MyAccountLayout
+        isRepresentative={isRepresentative}
+        accountName={accountName}
+      >
         <BeforeSection />
         <MyAccountListOrders
           listOrders={listOrders}
@@ -129,6 +135,11 @@ export const getServerSideProps: GetServerSideProps<
       },
     }
   }
+
+  const isRepresentative = getIsRepresentative({
+    headers: context.req.headers as Record<string, string>,
+    account: storeConfig.api.storeId,
+  })
 
   const { previewData } = context
 
@@ -233,6 +244,7 @@ export const getServerSideProps: GetServerSideProps<
         text,
         clientEmail,
       },
+      isRepresentative,
     },
   }
 }