+++ packages/core/src/pages/account/profile.tsx
@@ -9,12 +9,14 @@ import { MyAccountLayout } from 'src/components/account'
 import RenderSections from 'src/components/cms/RenderSections'
 import { default as GLOBAL_COMPONENTS } from 'src/components/cms/global/Components'
 import CUSTOM_COMPONENTS from 'src/customizations/src/components'
+import { ProfileSection } from 'src/components/account/profile'
 
 import { getGlobalSectionsData } from 'src/components/cms/GlobalSections'
 
 import { default as AfterSection } from 'src/customizations/src/myAccount/extensions/profile/after'
 import { default as BeforeSection } from 'src/customizations/src/myAccount/extensions/profile/before'
 import type { MyAccountProps } from 'src/experimental/myAccountSeverSideProps'
+import { getIsRepresentative } from 'src/sdk/account/getIsRepresentative'
 import { injectGlobalSections } from 'src/server/cms/global'
 import { getMyAccountRedirect } from 'src/utils/myAccountRedirect'
 import { gql } from '@generated/gql'
@@ -24,6 +26,7 @@ import type {
   ServerProfileQueryQueryVariables,
 } from '@generated/graphql'
 import { validateUser } from 'src/sdk/account/validateUser'
+import storeConfig from '../../../discovery.config'
 
 /* A list of components that can be used in the CMS. */
 const COMPONENTS: Record<string, ComponentType<any>> = {
@@ -31,10 +34,20 @@ const COMPONENTS: Record<string, ComponentType<any>> = {
   ...CUSTOM_COMPONENTS,
 }
 
+type ProfilePagePros = {
+  accountProfile: {
+    name: string | null
+    email: string | null
+    id: string | null
+  }
+} & MyAccountProps
+
 export default function Profile({
   globalSections,
   accountName,
-}: MyAccountProps) {
+  accountProfile,
+  isRepresentative,
+}: ProfilePagePros) {
   return (
     <RenderSections
       globalSections={globalSections.sections}
@@ -42,11 +55,12 @@ export default function Profile({
     >
       <NextSeo noindex nofollow />
 
-      <MyAccountLayout accountName={accountName}>
+      <MyAccountLayout
+        isRepresentative={isRepresentative}
+        accountName={accountName}
+      >
         <BeforeSection />
-        <div>
-          <h1>Profile</h1>
-        </div>
+        <ProfileSection profile={accountProfile} />
         <AfterSection />
       </MyAccountLayout>
     </RenderSections>
@@ -56,6 +70,11 @@ export default function Profile({
 const query = gql(`
   query ServerProfileQuery {
     accountName
+    accountProfile {
+      name
+      email
+      id
+    }
   }
 `)
 
@@ -74,6 +93,11 @@ export const getServerSideProps: GetServerSideProps<
       },
     }
   }
+
+  const isRepresentative = getIsRepresentative({
+    headers: context.req.headers as Record<string, string>,
+    account: storeConfig.api.storeId,
+  })
 
   const { isFaststoreMyAccountEnabled, redirect } = getMyAccountRedirect({
     query: context.query,
@@ -126,6 +150,8 @@ export const getServerSideProps: GetServerSideProps<
     props: {
       globalSections: globalSectionsResult,
       accountName: profile.data.accountName,
+      accountProfile: profile.data.accountProfile,
+      isRepresentative,
     },
   }
 }