+++ packages/core/src/pages/account/security.tsx
@@ -16,6 +16,7 @@ import { getGlobalSectionsData } from 'src/components/cms/GlobalSections'
 import { default as AfterSection } from 'src/customizations/src/myAccount/extensions/security/after'
 import { default as BeforeSection } from 'src/customizations/src/myAccount/extensions/security/before'
 import type { MyAccountProps } from 'src/experimental/myAccountSeverSideProps'
+import { getIsRepresentative } from 'src/sdk/account/getIsRepresentative'
 import { injectGlobalSections } from 'src/server/cms/global'
 import { getMyAccountRedirect } from 'src/utils/myAccountRedirect'
 import { execute } from 'src/server'
@@ -25,6 +26,8 @@ import type {
   ServerSecurityQueryQueryVariables,
 } from '@generated/graphql'
 import { validateUser } from 'src/sdk/account/validateUser'
+import storeConfig from '../../../discovery.config'
+import { SecuritySection } from 'src/components/account/security'
 
 /* A list of components that can be used in the CMS. */
 const COMPONENTS: Record<string, ComponentType<any>> = {
@@ -39,6 +42,7 @@ type SecurityPageProps = {
 export default function Page({
   globalSections,
   accountName,
+  isRepresentative,
 }: SecurityPageProps) {
   return (
     <RenderSections
@@ -47,11 +51,12 @@ export default function Page({
     >
       <NextSeo noindex nofollow />
 
-      <MyAccountLayout accountName={accountName}>
+      <MyAccountLayout
+        isRepresentative={isRepresentative}
+        accountName={accountName}
+      >
         <BeforeSection />
-        <div>
-          <h1>Security</h1>
-        </div>
+        <SecuritySection />
         <AfterSection />
       </MyAccountLayout>
     </RenderSections>
@@ -79,6 +84,11 @@ export const getServerSideProps: GetServerSideProps<
       },
     }
   }
+
+  const isRepresentative = getIsRepresentative({
+    headers: context.req.headers as Record<string, string>,
+    account: storeConfig.api.storeId,
+  })
 
   const { isFaststoreMyAccountEnabled, redirect } = getMyAccountRedirect({
     query: context.query,
@@ -131,6 +141,7 @@ export const getServerSideProps: GetServerSideProps<
     props: {
       globalSections: globalSectionsResult,
       accountName: security.data.accountName,
+      isRepresentative,
     },
   }
 }