+++ packages/core/test/server/index.test.ts
@@ -76,6 +76,8 @@ const QUERIES = [
   'productCount',
   'userOrder',
   'listUserOrders',
+  'userDetails',
+  'accountProfile',
   'accountName',
   'validateUser',
 ]
@@ -85,6 +87,8 @@ const MUTATIONS = [
   'validateSession',
   'subscribeToNewsletter',
   'cancelOrder',
+  'processOrderAuthorization',
+  'setPassword',
 ]
 
 describe('FastStore GraphQL Layer', () => {